requirements.txt
streamlit
pandas
numpy
openai
python-docx
Pillow
openpyxl
python-dotenv
xlsxwriter